requirements.txt
langchain
langchain-community
langchain-openai
langgraph
langchain_tavily
python-dotenv
langchain-groq
langchain-chroma
langchain-google-genai
langchain-text-splitters
pillow